Added 20
Row 6 must contain 2 reverse squares. We are restricted to 2 digit
squares because we know row 6 contains no actual squares. 61 cannot be
in this row as we know 60 is in row 1 and it is not possible to
traverse 5 rows in one move. 52 cannot be in row 6 as all multiple of
13 are accounted for in other rows. 63 cannot be in row 6 because it is
a multiple of row 1 and all multiples of 21 are accounted for in other
rows except for 1 which is an even number whereas 63 is odd -  this
means that 46 and 18 must both be in row 6. Multiples of 20 can only be
in rows 1 and 3, given that we have proven that 18 is in row 6 and we
know it is not possible to get to row 1 from 18 in 2 moves, 20 must
therefore be in row 4. We know that there is only one multiple of 20 in
row 4 and that location is F4 therefore this is the location of 20.