Added 40
21 must either be in col B or row 3 because this is where the 21
multiples are located. It cannot be in column B because you cannot get
to the location of 20 in one move. Therefore it must be somewhere else
in row 3.
Therefore the remaining multiple of 21 i.e. 42 must be in column B. We
knew that 40 was either in 1c or 1g as it must be on an even square in
row 1 however G1 is too far away from col B to get there in 2 moves
from cis B therefore it must be in C1.