Added 24
The even numbers in row 6 add up to 116, we know two of them are the
reverse squares (46 and 18). The remaining valid combinations of even
numbers is only : 2 and 50 - everything else doesn’t fit. This means 24
cannot go in row 6 it must therefore be in F8 as we know it must be one
move away from 25 and be in column F (because of the 3 multiples of 4 =
76 clue).