Added 4
Row must contain primes 3 and 5 because it must contain 2 more primes
below 20 to satisfy the 5 below 20 rule with whats there already. These
are the only candidates that are suitable. 10 cannot go in D6 because
all the even numbers in row 6 are 48, 18, 30 and 2 so 11 is not
reachable from 9 in 2 moves. If 5 and 3 are both in row 8 then 4 must
be in E7 as that is the only permittable move.