Added 37 and 38
G1 has only one blank space a knights move away, this means that it
must be consecutive to one of the numbers a knights move away. The
options are 21 (for which both 20 and 22 are both on the board) and 39
(for which 40 is already on the board). This makes G1 38 and 37 is
therefore in the blank space, one knights move from 38